Updated Code
Tested all this code on my own account because it took a while to update the code on github's end. The majority of this code should work or all of it should work now.  I wanted to update the code so github can then start to load this new code at least.